Move download function to their own module.
