+++ src/routes/power-cards/index.svelte
@@ -2,6 +2,7 @@
   import { onMount } from "svelte";
 
   import * as Lib from "../lib";
+  import { downloadHTML } from "$lib/download";
   import PreviewFrame from "$lib/preview-frame/index.svelte";
   import LoadButton from "$lib/load-button.svelte";
 
@@ -188,7 +189,7 @@
 
   function exportPowerCards() {
     const htmlFileName = powerCards.spiritName.replaceAll(" ", "_") + "_PowerCards.html";
-    Lib.downloadHTML(generateHTML(powerCards), htmlFileName);
+    downloadHTML(generateHTML(powerCards), htmlFileName);
   }
 
   function clearAllFields() {
